refactor: ViewModel classe shierarchy
